Revert "Pin mypy"
