+++ CHANGELOG.md
@@ -11,7 +11,8 @@
   internal infrastructure and unpublished data. It is now the responsibility of
   other programs utlizing this library to provide masks, input TBs, etc.
 * Replace `BootstrapAlgError` in `get_linfit` with a logged warning. Use default
-* Separate AMSR2 and AMSRE parameter coefficient specifications.
+  slope and offset values instead of failing when there are less than 125 valid
+  pixels.
 
 # v0.2.0
