+++ environment.yml
@@ -6,7 +6,7 @@ channels:
 dependencies:
   - invoke ~=1.7
   - isort ~=5.10
-  - mypy ==1.7.0
+  - mypy ~=1.5
   - pytest ~=7.1
   - pytest-cov ~=4.1.0
   - vulture ~=2.4
