--- pm_icecon/nt/params/__init__.py
@@ -1,3 +1,0 @@
-# Allow imports from `seaice_ecdr.nt.params.params` to be done as
-# `seaice_ecdr.nt.params`
-from .params import *
