+++ pyproject.toml
@@ -20,21 +20,6 @@ target-version = ["py310"]
 paths = ["pm_icecon"]
 min_confidence = 65
 
-[tool.ruff]
-target-version = "py310"
-
-extend-fixable = ["I"]
-extend-select = ["I"]
-
-ignore = [
-  # D1: Ignore errors requiring docstrings on everything.
-  # D203: "1 blank line required before class docstring"
-  # D213: "Multi-line docstring summary should start at the second line"
-  # E731: Lambda assignments are OK, use your best judgement.
-  # RUF010: !a, !r, !s are too magical for me.
-  "D1", "D203", "D213", "E731", "RUF010", "F403",
-]
-
 [tool.ruff.isort]
 known-first-party = ["pm_icecon"]
